+++ app/src/opencl_common.rs
@@ -5,6 +5,7 @@ use std::io::{Read, Write};
 use ocl::enums::{ProgramInfoResult, ProgramInfo};
 use ocl::{Buffer, Context, Device, EventList, Kernel, Platform, Program, Queue};
 
+#[allow(dead_code)]
 struct OpenCLResources {
     program: Program,
     queue: Queue,
@@ -15,11 +16,19 @@ struct OpenCLResources {
     buffer_best_hashes: Buffer::<u8>,
 }
 
-fn initialize_opencl(cnf: &PoWorkConf) -> Vec<OpenCLResources> {
-    if cnf.localsize != 256 {
+fn initialize_opencl(
+    diamond_mining: bool,
+    opencldir: &String,
+    platformid: &u32,
+    deviceids: &String,
+    workgroups: &u32,
+    localsize: &u32,
+    unitsize: &u32,
+) -> Vec<OpenCLResources> {
+    if *localsize != 256 {
         eprintln!(
             "[Warn] OpenCL local_size={} is incompatible with kernel fixed local arrays(256), fallback to CPU miner.",
-            cnf.localsize
+            localsize
         );
         return Vec::new();
     }
@@ -86,7 +95,7 @@ fn initialize_opencl(cnf: &PoWorkConf) -> Vec<OpenCLResources> {
         }
 
         let device_name = device.name().expect("Can't get device name");
-        let binary_file = format!(r"{}{}_{}{}.bin", opencldir, device_name, cnf_devices[idx], if diamond_mining { "_diamond" } else { "" });
+        let binary_file = format!(r"{}{}_{}{}.bin", opencldir, device_name, cnf_devices[idx], if diamond_mining { "_diamonds" } else { "" });
         let binary_path = Path::new(&binary_file);
 
         // Check if kernel was changed since last time (and need recompile)
@@ -223,71 +232,3 @@ fn compile_program_from_source(
     program
 }
 
-fn do_group_block_mining_opencl(
-    opencl: &OpenCLResources,
-    height: u64,
-    block_intro: Vec<u8>,
-    nonce_start: u32,
-    num_work_groups: u32,
-    local_work_size: u32,
-    unit_size: u32,
-) -> (u32, [u8; 32]) {
-    let mut most_nonce = 0u32;
-    let mut most_hash = [255u8; 32];
-    let global_work_size = num_work_groups * local_work_size;
-    let repeat = x16rs::block_hash_repeat(height) as u32;
-
-    let buffer_block_intro = Buffer::<u8>::builder()
-        .queue(opencl.queue.clone())
-        .flags(ocl::core::MEM_READ_ONLY)
-        .len(block_intro.len())
-        .copy_host_slice(&block_intro)
-        .build()
-        .expect("Unable to create buffer_block_intro");
-
-    let kernel = Kernel::builder()
-        .program(&opencl.program)
-        .name("x16rs_main")
-        .queue(opencl.queue.clone())
-        .global_work_size(global_work_size)
-        .local_work_size(local_work_size)
-        .arg(&buffer_block_intro)
-        .arg(nonce_start)
-        .arg(repeat)
-        .arg(unit_size)
-        .arg(&opencl.buffer_global_hashes)
-        .arg(&opencl.buffer_global_order)
-        .arg(&opencl.buffer_best_hashes)
-        .arg(&opencl.buffer_best_nonces)
-        .build()
-        .unwrap();
-
-    let mut kernel_event = EventList::new();
-    unsafe {
-        kernel.cmd().enew(&mut kernel_event).enq().expect("Unable to queue OpenCL kernel");
-    }
-
-    let mut hashes = vec![0u8; opencl.buffer_best_hashes.len()];
-    opencl.buffer_best_hashes
-        .read(&mut hashes)
-        .ewait(&kernel_event)
-        .enq()
-        .expect("Can't read buffer_best_hashes");
-
-    let mut nonces = vec![0u32; opencl.buffer_best_nonces.len()];
-    opencl.buffer_best_nonces
-        .read(&mut nonces)
-        .ewait(&kernel_event)
-        .enq()
-        .expect("Can't read buffer_best_nonces");
-
-    for i in 0..num_work_groups as usize {
-        let hash_bytes = &hashes[i * 32..(i * 32) + 32];
-        if hash_more_power(hash_bytes, &most_hash) {
-            most_hash.copy_from_slice(hash_bytes);
-            most_nonce = nonces[i];
-        }
-    }
-    
-    (most_nonce, most_hash)
-}
